% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alliances_1993}
\alias{alliances_1993}
\title{Correlates of War 1993: Alliances among nations}
\format{A network object:
\describe{
  \item{alliances_1993}{network}
}}
\source{
\url{http://www.correlatesofwar.org/}.
data(alliances_1993)
alliances_1993
}
\usage{
alliances_1993
}
\description{
Correlates of War 1993: Alliances among nations
}
\keyword{datasets}

