% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coleman}
\alias{coleman}
\title{Coleman's High School Friendship Data}
\format{An array with 2 matrices which are each 73 by 73:
\describe{
  \item{coleman}{array}
}}
\source{
Coleman, J. S. (1964). Introduction to Mathermatical Sociology. New York: Free Press.
}
\usage{
coleman
}
\description{
ames Coleman (1964) reports research on self-reported friendship ties
among 73 boys in a small high school in Illinois over the 1957-1958
academic year. Networks of reported ties for all 73 informants are
provided for two time points (fall and spring).
}
\details{
Both networks reflect answers to the question,
“What fellows here in school do you go around with most often?”
with the presence of an (i,j,k) edge indicating that j nominated
k in time period i.
 The data are unvalued and directed; although the self-reported ties
 are highly reciprocal, unreciprocated nominations are possible.
 It should be noted that, although this data is usually described as
 “friendship,” the sociometric item employed might be more accurately
 characterized as eliciting “frequent elective interaction.” This should
 be borne in mind when interpreting this data.
}
\examples{
data(coleman)
dim(coleman)
dimnames(coleman)[1]
}
\keyword{datasets}

