% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contig_1993}
\alias{contig_1993}
\title{Correlates of War 1993: Contiguity among nations}
\format{A network object:
\describe{
  \item{contig_1993}{network}
}}
\source{
\url{http://www.correlatesofwar.org/}.
data(contig_1993)
contig_1993
}
\usage{
contig_1993
}
\description{
Correlates of War 1993: Contiguity among nations
}
\keyword{datasets}

