% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kaptail.ins}
\alias{kaptail.ins}
\title{Kapferer–Tailor Shop}
\format{A network object:
\describe{
  \item{kaptail.ins}{network}
}}
\source{
Kapferer B. (1972). Strategy and transaction in an African factory. Manchester: Manchester University Press.
}
\usage{
kaptail.ins
}
\description{
Bruce Kapferer (1972) observed interactions in a tailor shop in
Zambia (then Northern Rhodesia) over a period of ten months.
His focus was the changing patterns of alliance among workers
during extended negotiations for higher wages.
}
\examples{
data(kaptail.ins)
kaptail.ins
}
\keyword{datasets}

