% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kfr}
\alias{kfr}
\title{Krackhardt Office Cognitive Social Structure}
\format{An array with 21 matrices which are each 21 by 21:
\describe{
  \item{kfr}{array}
}}
\source{
Krackhardt D. (1987). Cognitive social structures. Social Networks, 9, 104-134.
}
\usage{
kfr
}
\description{
David Krackhardt collected cognitive social structure data
from 21 management personnel in a high-tech, machine manufacturing
firm to assess the effects of a recent management intervention program.
The relation queried was “Who is a friend of X?" (krackfr). Each person indicated
not only his or her own advice and friendship relationships, but also
the relations he or she perceived among all other managers, generating a
full 21 by 21 matrix of adjacency ratings from each person in the group.
}
\examples{
data(kfr)
dim(kfr)
}
\keyword{datasets}

