% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coreFunctions.R
\name{lab1}
\alias{lab1}
\title{Lab 1}
\usage{
lab1()
}
\arguments{
\item{Empty}{}
}
\description{
\code{lab1} returns simple description of the first lab in Soc 88412.
}
\details{
This is a simple description of the lab, use help(lab1) to review the
examples for this lab
}
\examples{
## Practice R code
matrix(c(1,2,3,4),nc=2,nr=2)
## More to come!
\dontrun{

}
}

