% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mids_1993}
\alias{mids_1993}
\title{Correlates of War 1993: Militarized interstate disputes}
\format{A network object:
\describe{
  \item{mids_1993}{network}
}}
\source{
\url{http://www.correlatesofwar.org/}.
}
\usage{
mids_1993
}
\description{
Correlates of War 1993: Militarized interstate disputes
}
\examples{
data(mids_1993)
mids_1993
}
\keyword{datasets}

