% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nkad}
\alias{nkad}
\title{Krackhardt’s group of 21 high-tech managers, self-reported advice-seeking}
\format{A network object:
\describe{
  \item{nkad}{network}
}}
\source{
data(nkad)
nkad
}
\usage{
nkad
}
\description{
Krackhardt’s group of 21 high-tech managers, self-reported advice-seeking
}
\keyword{datasets}

