% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nkfr}
\alias{nkfr}
\title{Krackhardt’s group of 21 high-tech managers, self-reported friendship}
\format{A network object:
\describe{
  \item{nkfr}{network}
}}
\source{
data(nkfr)
nkfr
}
\usage{
nkfr
}
\description{
Krackhardt’s group of 21 high-tech managers, self-reported friendship
}
\keyword{datasets}

