% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nkrep}
\alias{nkrep}
\title{Krackhardt’s group of 21 high-tech managers, self-reported formal reporting}
\format{A network object:
\describe{
  \item{nkrep}{network}
}}
\source{
data(nkrep)
nkrep
}
\usage{
nkrep
}
\description{
Krackhardt’s group of 21 high-tech managers, self-reported formal reporting
}
\keyword{datasets}

