% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{silsys.fr.ilas}
\alias{silsys.fr.ilas}
\title{David Krackhardt's famous  Silicon Valley Systems study}
\format{A network object:
\describe{
  \item{silsys.fr.ilas}{network}
}}
\source{
data(silsys.fr.ilas)
silsys.fr.ilas
}
\usage{
silsys.fr.ilas
}
\description{
The two networks consist of friendship, as agreed upon by the ties' endpoints.
}
\keyword{datasets}

