% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sw.incidence}
\alias{sw.incidence}
\title{Southern Women}
\format{A matrix object:
\describe{
  \item{sw.incidence}{matrix}
}}
\source{
Davis, A., Gardner, B. B., and Gardner, M. R. (2009). Deep South: A social anthropological study of caste and class. Univ of South Carolina Press.
data(sw.incidence)
sw.incidence
}
\usage{
sw.incidence
}
\description{
This is the incidence matrix for the famous ``Southern Women" data set from Davis,
Gardner, and Gardner's Deep South study of class and social interaction in
Natchez, MS during the mid-1930s. The matrix shows the attendance of 18 women at
14 informal social events during a nine-month observation period, based on
``interviews, the records of participant observers, guest lists, and the newspapers" (DGG, p. 149).
}
\keyword{datasets}

