% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{thuroff.int}
\alias{thuroff.int}
\title{Thurman Office Network}
\format{A network object:
\describe{
  \item{thuroff.int}{network}
}}
\source{
Thurman B. (1979). In the office: Networks and coalitions. Social Networks, 2, 47-63.
data(thuroff.int)
thuroff.int
}
\usage{
thuroff.int
}
\description{
Thurman spent 16 months observing the interactions among
employees in the overseas office of a large international
corporation. During this time, two major disputes erupted
in a subgroup of fifteen people. Thurman analyzed the
outcome of these disputes in terms of the network of
formal and informal associations among those involved.
}
\keyword{datasets}

